refactor detail page: customer reviews
- separate form and review list
