+++ src/views/detail-view/restaurant-reviews/restaurant-reviews.scss
@@ -1,0 +1,100 @@
+@import 'src/styles/config';
+@import 'src/styles/helpers';
+
+restaurant-reviews {
+    .reviews {
+        &__form {
+            margin: 2rem 0;
+        }
+
+        &__title {
+            margin-top: 1rem;
+        }
+    }
+
+    .form {
+        &__footer {
+            width: 100%;
+            display: flex;
+            flex-direction: column;
+            margin: 0.5rem 0;
+        }
+
+        &__main {
+            width: 100%;
+        }
+
+        &__review, &__name {
+            flex: 1;
+            font-family: inherit;
+            font-size: 1rem;
+            padding: 1rem;
+            border: none;  
+            width: 100%;
+            outline: none;
+            border-bottom: 1px solid $medium-grey;
+
+            &:focus {
+                border-bottom: 2px solid $black;
+            }
+        }
+
+        &__name {
+            height: 44px;
+        }
+
+        &__review {
+            resize: none;
+            overflow: hidden;
+        }
+
+        &__buttons {
+            display: flex;
+            flex: 1;
+            flex-direction: row;
+            justify-content: flex-end;
+            padding: 0.5rem 0;
+        }
+
+        &__button {
+            margin-left: 0.5rem;
+            padding: 0.5rem 1rem;
+            min-height: 44px;
+            min-width: 44px;
+            text-transform: uppercase;
+            font-weight: 700;
+            font-size: 0.875rem;
+            border: none;
+            border-radius: 0.25rem;
+            cursor: pointer;
+
+            
+            &.submit {
+                background: $color-rating-star;
+                &:hover {
+                    background: 
+                        lighten(
+                            $color:$color-rating-star, 
+                            $amount: 5%
+                        );
+                }
+            }
+
+            &.cancel {
+                background: $color-rating-base;
+                &:hover {
+                    background: 
+                        lighten(
+                            $color:$color-rating-base, 
+                            $amount: 5%
+                        );
+                }
+            }
+        }
+    }
+
+    .divider {
+        border: 0.5px solid $color-rating-star;
+        margin: 0.5rem 0;
+    }
+}
