+++ src/views/detail-view/restaurant-reviews/review-list/review-list.scss
@@ -1,0 +1,38 @@
+@import 'src/styles/config';
+@import 'src/styles/helpers';
+
+review-list {
+    display: grid;
+    gap: 2rem;
+    
+    .review {
+        display: grid;
+        grid-template-columns: 3.5rem auto;
+
+        &__author-thumbnail {
+            i {
+                font-size: 2.5rem;
+            }
+        }
+
+        &__main {
+            display: grid;
+            gap: 1rem;
+        }
+
+        &__header {
+            @include flex-center-vertical(row);
+        }
+        
+        &__author {
+            font-weight: 700;
+            margin-right: 0.5rem;
+        }
+
+        &__date {
+            font-size: 0.875rem;
+            color: $color-text-subtle;
+        }
+
+    }
+}
