+++ src/views/detail-view/restaurant-reviews/review-list/review-list_responsive.scss
@@ -1,0 +1,7 @@
+@import 'src/styles/config';
+
+review-list {
+    @include responsive-min($medium) {
+        grid-template-columns: repeat(2, 1fr);
+    }
+}
